remove swg
